+++ src/Symfony/Component/Validator/Constraints/IsNull.php
@@ -19,6 +19,7 @@ use Symfony\Component\Validator\Constraint;
  *
  * @author Bernhard Schussek <bschussek@gmail.com>
  */
+#[\Attribute(\Attribute::TARGET_PROPERTY | \Attribute::TARGET_METHOD | \Attribute::IS_REPEATABLE)]
 class IsNull extends Constraint
 {
     const NOT_NULL_ERROR = '60d2f30b-8cfa-4372-b155-9656634de120';
@@ -28,4 +29,11 @@ class IsNull extends Constraint
     ];
 
     public $message = 'This value should be null.';
+
+    public function __construct(array $options = null, string $message = null, array $groups = null, $payload = null)
+    {
+        parent::__construct($options ?? [], $groups, $payload);
+
+        $this->message = $message ?? $this->message;
+    }
 }