+++ src/Symfony/Component/Validator/Tests/Fixtures/Annotation/GroupSequenceProviderEntity.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Component\Validator\Tests\Fixtures;
+namespace Symfony\Component\Validator\Tests\Fixtures\Annotation;
 
 use Symfony\Component\Validator\Constraints as Assert;
 use Symfony\Component\Validator\GroupSequenceProviderInterface;