+++ src/Symfony/Component/Validator/CHANGELOG.md
@@ -32,6 +32,7 @@ CHANGELOG
  * added the `Isin` constraint and validator
  * added the `ULID` constraint and validator
  * added support for UUIDv6 in `Uuid` constraint
+ * enabled the validator to load constraints from PHP attributes
 
 5.1.0
 -----